Fixed tests

+++ src/test/java/inc/evil/aws/fiddle/common/DynamoDbCsvTestExecutionListener.java
@@ -1,6 +1,8 @@
 package inc.evil.aws.fiddle.common;
 
 import java.io.IOException;
+import java.util.Arrays;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import com.fasterxml.jackson.databind.MappingIterator;
 import com.fasterxml.jackson.dataformat.csv.CsvMapper;
@@ -15,36 +17,55 @@ import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
 
 public class DynamoDbCsvTestExecutionListener implements TestExecutionListener {
+    private final AtomicBoolean cleanupFlag = new AtomicBoolean();
 
     @Override
-    public void beforeTestMethod(TestContext testContext) throws Exception {
+    public void beforeTestMethod(TestContext testContext) {
         ApplicationContext applicationContext = testContext.getApplicationContext();
-        DynamoCsv dynamoCsv = testContext.getTestMethod().getAnnotation(DynamoCsv.class);
+        DynamoCsv[] annotations = testContext.getTestMethod().getAnnotationsByType(DynamoCsv.class);
+        boolean shouldCleanupTable = shouldCleanupTable(annotations);
+        for (DynamoCsv dynamoCsv : annotations) {
+            executeFor(applicationContext, dynamoCsv, shouldCleanupTable);
+        }
+    }
+
+    private boolean shouldCleanupTable(DynamoCsv[] annotations) {
+        return Arrays.stream(annotations)
+                     .map(DynamoCsv::cleanupTable)
+                     .reduce(false, (a, b) -> a | b);
+    }
+
+    private void executeFor(ApplicationContext applicationContext, DynamoCsv dynamoCsv, boolean shouldCleanupTable) {
         String resourcePath = dynamoCsv.value();
         Class<?> entityType = dynamoCsv.entityType();
 
         DynamoDbTable<Object> table = makeDynamoDbTable(applicationContext, entityType);
 
-        if (dynamoCsv.reCreateTable()) {
-            dropAndCreateTable(table);
+        if (shouldCleanupTable) {
+            cleanupTable(table);
         }
+
         if (!resourcePath.isEmpty()) {
             readAndPutItems(applicationContext, resourcePath, entityType, table);
         }
     }
 
-    private static void readAndPutItems(ApplicationContext applicationContext, String resourcePath, Class<?> entityType, DynamoDbTable<Object> table)
-        throws IOException {
+    private static void readAndPutItems(ApplicationContext applicationContext, String resourcePath, Class<?> entityType, DynamoDbTable<Object> table) {
         try (MappingIterator<Object> mappingIterator = parseCvs(entityType, applicationContext.getResource(resourcePath))) {
             mappingIterator.readAll().forEach(table::putItem);
+        } catch (IOException e) {
+            throw new RuntimeException();
         }
     }
 
-    private static void dropAndCreateTable(DynamoDbTable<Object> table) {
-        table.scan()
-             .stream()
-             .flatMap(page -> page.items().stream())
-             .forEach(table::deleteItem);
+    private void cleanupTable(DynamoDbTable<Object> table) {
+        if (!cleanupFlag.get()) {
+            table.scan()
+                 .stream()
+                 .flatMap(page -> page.items().stream())
+                 .forEach(table::deleteItem);
+        }
+        cleanupFlag.compareAndSet(false, true);
     }
 
     @SuppressWarnings("unchecked")
